# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /api/v1/users/me/category-rules:
    post:
      description: Matching normalizes surrounding and repeated whitespace and compares case-insensitively.
      operationId: CategoryRules_createCategoryRule
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRuleDto'
      responses:
        '201':
          description: Category rule created.
          headers:
            Location:
              required: true
              description: Relative canonical URI of the created Category rule.
              schema:
                type: string
                example: /api/v1/users/me/category-rules/42
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRuleResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Create a category rule.
      tags:
        - Category rules
    get:
      operationId: CategoryRules_listCategoryRules
      parameters: []
      responses:
        '200':
          description: Category rules.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryRuleResponseDto'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: List all owned category rules in ascending ID order.
      tags:
        - Category rules
  /api/v1/users/me/category-rules/{ruleId}:
    get:
      operationId: CategoryRules_getCategoryRule
      parameters:
        - name: ruleId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      responses:
        '200':
          description: Category rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRuleResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Get a category rule.
      tags:
        - Category rules
    patch:
      description: A category rule may be reassigned only to an active Category; assigning it to an inactive Category returns 409 Conflict.
      operationId: CategoryRules_updateCategoryRule
      parameters:
        - name: ruleId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRuleDto'
      responses:
        '200':
          description: Category rule updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRuleResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Update or reassign a category rule.
      tags:
        - Category rules
    delete:
      operationId: CategoryRules_deleteCategoryRule
      parameters:
        - name: if-match
          required: false
          in: header
          description: Optional Category Rule updatedAt timestamp. The delete is rejected when it is stale.
          schema:
            type: string
            format: date-time
        - name: ruleId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      responses:
        '204':
          description: Category rule deleted.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Delete a category rule.
      tags:
        - Category rules
...
  /api/v1/users/me/spaces/{spaceId}/category-rules/{ruleId}:
    get:
      operationId: SpaceCategoryRules_getCategoryRule
      parameters:
        - name: ruleId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
      responses:
        '200':
          description: Category rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRuleResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Get a Category Rule in an authorized Space.
      tags:
        - Category rules
    patch:
      description: The optional updatedAt timestamp rejects a stale edit. A Category Rule may be reassigned only to an active Category in the same Space.
      operationId: SpaceCategoryRules_updateCategoryRule
      parameters:
        - name: ruleId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRuleDto'
      responses:
        '200':
          description: Category rule updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRuleResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Update or reassign a Category Rule in an authorized Space.
      tags:
        - Category rules
    delete:
      operationId: SpaceCategoryRules_deleteCategoryRule
      parameters:
        - name: if-match
          required: false
          in: header
          description: Optional Category Rule updatedAt timestamp. The delete is rejected when it is stale.
          schema:
            type: string
            format: date-time
        - name: ruleId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
      responses:
        '204':
          description: Category rule deleted.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Delete a Category Rule in an authorized Space.
      tags:
        - Category rules
...
info:
  title: Expense Tracker REST API
  description: A user-scoped expense tracker API. Bigint identifiers and counts, decimal money, domain dates, and UTC timestamps use the encodings described by the schemas below.
  version: 1.0.0
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - INVALID_JSON
        - NOT_ACCEPTABLE
        - UNSUPPORTED_MEDIA_TYPE
        - ROUTE_NOT_FOUND
        - HTTP_ERROR
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
        - UNAUTHENTICATED
        - USER_NOT_PROVISIONED
        - CONFLICT
        - STALE_EDIT
        - RESOURCE_NOT_FOUND
        - USER_NOT_FOUND
        - SPACE_NOT_FOUND
        - SPACE_NOT_WRITABLE
        - EMAIL_ALREADY_EXISTS
        - CATEGORY_NOT_FOUND
        - CATEGORY_NAME_ALREADY_EXISTS
        - CATEGORY_INACTIVE
        - BUDGET_NOT_FOUND
        - CATEGORY_RULE_NOT_FOUND
        - CATEGORY_RULE_PATTERN_ALREADY_EXISTS
        - TRANSACTION_NOT_FOUND
        - STATEMENT_IMPORT_NOT_FOUND
        - IMPORTED_TRANSACTION_IMMUTABLE
        - STATEMENT_IMPORT_FILE_ALREADY_EXISTS
        - STATEMENT_IMPORT_FILE_HASH_INVALID
        - STATEMENT_IMPORT_PROBABLE_DUPLICATES
      description: Stable error code. Human-readable messages are not a compatibility contract.
    ErrorResponseDto:
      type: object
      properties:
        code:
          description: Stable error code. Human-readable messages are not a compatibility contract.
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable error message.
        details:
          type: array
          description: Structured details for fields or domain-specific conflicts.
          items:
            oneOf:
              - $ref: '#/components/schemas/ErrorDetailDto'
              - $ref: '#/components/schemas/ProbableDuplicateDetailDto'
      additionalProperties: false
      required:
        - code
        - message
        - details
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponseDto'
      additionalProperties: false
      required:
        - error
    ErrorDetailCode:
      type: string
      enum:
        - required
        - invalid_format
        - invalid_type
        - invalid_value
        - unknown
        - incompatible
        - not_allowed
        - not_unique
        - too_short
        - too_long
        - probable_duplicate
      description: Stable machine-readable field-detail code.
    ErrorDetailDto:
      type: object
      properties:
        categoryId:
          type: string
          description: Conflicting owned Category identifier for a category-rule duplicate.
          pattern: ^[1-9]\d*$
          example: '42'
        field:
          type: string
          description: JSON Pointer into the request.
          example: /email
        code:
          description: Stable machine-readable field-detail code.
          allOf:
            - $ref: '#/components/schemas/ErrorDetailCode'
        message:
          type: string
          description: Human-readable detail message.
      additionalProperties: false
      required:
        - field
        - code
        - message
    ProbableDuplicateDetailDto:
      type: object
      properties:
        field:
          type: string
          example: /transactions/0
        code:
          type: string
          enum:
            - probable_duplicate
        message:
          type: string
        transactionIndexes:
          type: array
          items:
            type: integer
            minimum: 0
        committedTransactionIds:
          type: array
          description: Public transaction IDs encoded as strings.
          items:
            type: string
            pattern: ^[1-9]\d*$
      additionalProperties: false
      required:
        - field
        - code
        - message
        - transactionIndexes
        - committedTransactionIds
...
    CategoryRuleResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Positive bigint Category rule identifier encoded as a string.
          pattern: ^[1-9]\d*$
          example: '42'
        categoryId:
          type: string
          description: Positive bigint Category identifier encoded as a string.
          pattern: ^[1-9]\d*$
          example: '42'
        pattern:
          type: string
          description: Original rule text containing at least one non-whitespace character. Matching trims surrounding whitespace, collapses repeated whitespace, and compares case-insensitively.
          minLength: 1
          maxLength: 500
          pattern: \S
          example: Green   Market
        matchType:
          type: string
          description: Exact or Contains matching. The normalized pattern is not returned.
          enum:
            - exact
            - contains
          example: exact
        createdAt:
          type: string
          description: UTC timestamp when the Category rule was created.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
        updatedAt:
          type: string
          description: UTC timestamp when the Category rule was last updated.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
      description: A Category Rule belonging to the authorized Space. The original pattern is retained for display. Rule evaluation is performed by the frontend.
      additionalProperties: false
      required:
        - id
        - categoryId
        - pattern
        - matchType
        - createdAt
        - updatedAt
    CreateCategoryRuleDto:
      type: object
      properties:
        matchType:
          type: string
          enum:
            - exact
            - contains
          default: exact
          description: Omission creates an Exact rule, including frontend Remember requests.
        pattern:
          type: string
          description: Original rule text. Uniqueness trims surrounding whitespace, collapses repeated whitespace, and ignores case within each match type. Rule evaluation belongs to the frontend.
          minLength: 1
          maxLength: 500
          pattern: \S
          example: Green Market
        categoryId:
          type: string
          description: Positive bigint Category identifier encoded as a decimal JSON string. The Category must be active in the destination Space; the legacy personal route resolves the authenticated User's Personal Space. An inactive Category causes a 409 Conflict.
          pattern: ^[1-9]\d*$
          example: '42'
      additionalProperties: false
      required:
        - pattern
        - categoryId
    UpdateCategoryRuleDto:
      type: object
      properties:
        matchType:
          type: string
          enum:
            - exact
            - contains
          description: Omission preserves the current match type.
        pattern:
          type: string
          description: Original rule text. Uniqueness trims surrounding whitespace, collapses repeated whitespace, and ignores case within each match type.
          minLength: 1
          maxLength: 500
          pattern: \S
          example: Green Market
        categoryId:
          type: string
          description: Positive bigint Category identifier encoded as a decimal JSON string. Reassignment requires an active Category in the destination Space; an inactive Category causes a conflict.
          pattern: ^[1-9]\d*$
          example: '42'
        updatedAt:
          type: string
          description: The UTC timestamp returned by the last read. A stale scoped edit is rejected when another member changed this rule first.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
      additionalProperties: false
      minProperties: 1
...
  responses:
    UnauthenticatedError:
      description: The request does not include valid Clerk session credentials.
      headers:
        WWW-Authenticate:
          required: true
          description: Authentication scheme accepted by the API.
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication required
              details: []
    UserNotProvisionedError:
      description: The verified Clerk identity is not linked to a local User. Provision the User first.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: USER_NOT_PROVISIONED
              message: User is not provisioned
              details: []
    NotFoundError:
      description: The requested resource is absent or owned by another user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            user:
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User was not found
                  details: []
            route:
              value:
                error:
                  code: ROUTE_NOT_FOUND
                  message: Request failed
                  details: []
    NotAcceptableError:
      description: The client did not accept the representation required by the endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: NOT_ACCEPTABLE
              message: The endpoint requires an application/json response
              details: []
    InternalError:
      description: An unexpected server error. Infrastructure details are never returned.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: INTERNAL_ERROR
              message: An internal error occurred
              details: []
    ValidationError:
      description: The request syntax or fields are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            invalidField:
              value:
                error:
                  code: VALIDATION_FAILED
                  message: The request contains invalid fields.
                  details:
                    - field: /email
                      code: invalid_format
                      message: Email must be a valid email address
    ConflictError:
      description: The request conflicts with current application state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            emailAlreadyExists:
              value:
                error:
                  code: EMAIL_ALREADY_EXISTS
                  message: Email is already in use
                  details:
                    - field: /email
                      code: not_unique
                      message: Email is already in use
            genericConflict:
              value:
                error:
                  code: CONFLICT
                  message: Request conflicts with existing data
                  details: []
...
    HttpError:
      description: A generic HTTP protocol error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: HTTP_ERROR
              message: Request failed
              details: []
    UnsupportedMediaTypeError:
      description: A request body must use application/json.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: Request bodies must use application/json
              details: []
...
security:
  - bearerAuth: []